added a version with list and cahgned some shit
I think we're going to have to do this by dictionary or get the other
way to work.

the dict will be like this:

btn_id, x, y

so when you call upon the btn_id, you can get the x and y from that. ill
get the logic on a sheet and a doy te manana