Fixed the error, but a new speculation has arisen
For the prefabs that have two solids, such as the wall prefabs that have
both the ground and the wall. there are two "world_id_num"s in the txt
file, so when the prefab code replaces them, they become the same exact
world id num in the vmf file. However, this does not affect the vmf
opening in hammer for some reason.