kinda fixed parentname again
I have a weird error. When I put only the spawnrooms, the parent works
fine, and the visualizer( the little (/) on the door) is in the correct
place. When I put spawnrooms AND other solids, the parent works fine
(when I look in the file), but the visualizer is off. From my
understanding, the visualizer works when the parent works. Before, when
the parent still was "parent_name", the visualizer didn't work, so...